Clean library
