+++ CyberSwift/Managers/RestAPIManager/RestAPIManager.swift
@@ -22,115 +22,7 @@ public class RestAPIManager {
     // MARK: - Properties
     public static let instance = RestAPIManager()
     
-    
-    // MARK: - Class Initialization
-    private init() {}
-    
-    deinit {
-        Logger.log(message: "Success", event: .severe)
-    }
-    
-    
-    // MARK: - Class Functions
-    public func generateKeys(userID: String, password: String) -> [String: UserKeys] {
-        var userKeys = [String: UserKeys]()
-        
-        // type
-        let types = ["owner", "active", "posting", "memo"]
-        
-        for keyType in types {
-            let seed                =   userID + keyType + password
-            let brainKey            =   seed.removeWhitespaceCharacters()
-            let brainKeyBytes       =   brainKey.bytes
-            
-            var brainKeyBytesSha256 =   brainKeyBytes.sha256()
-            brainKeyBytesSha256.insert(0x80, at: 0)
-            
-            let checksumSha256Bytes =   brainKeyBytesSha256.generateChecksumSha256()
-            brainKeyBytesSha256     +=  checksumSha256Bytes
-            
-            if let privateKey = PrivateKey(brainKeyBytesSha256.base58EncodedString) {
-                let publicKey = privateKey.createPublic(prefix: PublicKey.AddressPrefix.mainNet)
-                userKeys[keyType] = UserKeys(privateKey: privateKey.description, publicKey: publicKey.description)
-            }
-        }
-        
-        return userKeys
-    }
-    
-    
     // MARK: - FACADE-SERVICE
-    // API `auth.authorize`
-    public func authorize(userID:               String,
-                          userActiveKey:        String,
-                          responseHandling:     @escaping (ResponseAPIAuthAuthorize) -> Void,
-                          errorHandling:        @escaping (Error) -> Void) {
-        // Offline mode
-        if (!Config.isNetworkAvailable) { return errorHandling(ErrorAPI.disableInternetConnection(message: nil)) }
-        
-        let methodAPIType = MethodAPIType.authorize(userID: userID, activeKey: userActiveKey)
-
-        Broadcast.instance.executeGETRequest(
-            byContentAPIType: methodAPIType,
-            onResult: { responseAPIResult in
-                guard let result = (responseAPIResult as! ResponseAPIAuthAuthorizeResult).result
-                else {
-                    let responseAPIError = (responseAPIResult as! ResponseAPIAuthAuthorizeResult).error
-                    
-                    Logger.log(message: "\nAPI `auth.authorize` response mapping error: \n\(responseAPIError!.message)\n", event: .error)
-                    
-                    return errorHandling(ErrorAPI.jsonParsingFailure(message: "\(responseAPIError!.message)"))
-                }
-                
-                DispatchQueue.main.async(execute: {
-                    // Log result
-                    Logger.log(message: "\nAPI `auth.authorize` response result: \n\(responseAPIResult)\n", event: .debug)
-                    
-                    // Save to UserDefault
-                    UserDefaults.standard.set(true, forKey: Config.isCurrentUserLoggedKey)
-                    
-                    // Save in Keychain
-                    do {
-                        try KeychainManager.deleteUser()
-                        try KeychainManager.save(data: [
-                            Config.currentUserIDKey:            userID,
-                            Config.currentUserNameKey:          result.displayName,
-                            Config.currentUserPublicActiveKey:  userActiveKey
-                        ])
-                        
-                        // Save in iCloud key-value
-                        let keyStore = NSUbiquitousKeyValueStore()
-                        keyStore.set(userID, forKey: Config.currentUserIDKey)
-                        keyStore.set(result.displayName, forKey: Config.currentUserNameKey)
-                        keyStore.set(userActiveKey, forKey: Config.currentUserPublicActiveKey)
-                        keyStore.synchronize()
-                        
-                        // API `push.notifyOn`
-                        if let fcmToken = UserDefaults.standard.value(forKey: "fcmToken") as? String {
-                            RestAPIManager.instance.pushNotifyOn(
-                                fcmToken:          fcmToken,
-                                responseHandling:  { response in
-                                    Logger.log(message: response.status, event: .severe)
-                            },
-                                errorHandling:     { errorAPI in
-                                    Logger.log(message: errorAPI.caseInfo.message, event: .error)
-                            })
-                        }
-                        
-                        responseHandling(result)
-                        
-                    } catch {
-                        errorHandling(error)
-                        return
-                    }
-                })
-        },
-            onError:           { (errorAPI) in
-                Logger.log(message: "\nAPI `auth.authorize` response error: \n\(errorAPI.localizedDescription)\n", event: .error)
-                errorHandling(errorAPI)
-        })
-    }
-    
     // API `auth.generateSecret`
     private func generateSecret(completion: @escaping (ResponseAPIAuthGenerateSecret?, ErrorAPI?) -> Void) {
         // Offline mode
@@ -323,7 +215,7 @@ public class RestAPIManager {
     }
     
     // API `push.notifyOn`
-    private func pushNotifyOn(fcmToken:             String,
+    public func pushNotifyOn(fcmToken:             String,
                               responseHandling:     @escaping (ResponseAPINotifyPushOn) -> Void,
                               errorHandling:        @escaping (ErrorAPI) -> Void) {
         // Offline mode
@@ -710,247 +602,6 @@ public class RestAPIManager {
         })
     }
 
-
-    // MARK: - REGISTRATION-SERVICE
-    // API `registration.getState`
-    public func getState(id:                String? = Config.currentUser?.id,
-                         phone:             String?,
-                         responseHandling:  @escaping (ResponseAPIRegistrationGetState) -> Void,
-                         errorHandling:     @escaping (ErrorAPI) -> Void) {
-        // Offline mode
-        if (!Config.isNetworkAvailable) { return errorHandling(ErrorAPI.disableInternetConnection(message: nil)) }
-        
-        let methodAPIType = MethodAPIType.getState(id: id, phone: phone)
-
-        Broadcast.instance.executeGETRequest(byContentAPIType:  methodAPIType,
-                                             onResult:          { responseAPIResult in
-                                                guard let result = (responseAPIResult as! ResponseAPIRegistrationGetStateResult).result else {
-                                                    let responseAPIError = (responseAPIResult as! ResponseAPIRegistrationGetStateResult).error
-                                                    Logger.log(message: "\nAPI `registration.getState` response mapping error: \n\(responseAPIError!.message)\n", event: .error)
-                                                    return errorHandling(ErrorAPI.jsonParsingFailure(message: "\(responseAPIError!.message)"))
-                                                }
-                                                
-                                                Logger.log(message: "\nAPI `registration.getState` response result: \n\(responseAPIResult)\n", event: .debug)
-                                                responseHandling(result)
-        },
-                                             onError:           { errorAPI in
-                                                Logger.log(message: "\nAPI `registration.getState` response error: \n\(errorAPI.localizedDescription)\n", event: .error)
-                                                errorHandling(errorAPI)
-        })
-    }
-    
-    // API `registration.firstStep`
-    public func firstStep(phone:                String,
-                          responseHandling:     @escaping (ResponseAPIRegistrationFirstStep) -> Void,
-                          errorHandling:        @escaping (ResponseAPIError) -> Void) {
-        // Offline mode
-        if (!Config.isNetworkAvailable) { return errorHandling(ResponseAPIError(code: 503, message: "No Internet Connection", currentState: nil)) }
-        
-        let methodAPIType = MethodAPIType.firstStep(phone: phone, isDebugMode: isDebugMode)
-        
-        Broadcast.instance.executeGETRequest(byContentAPIType:  methodAPIType,
-                                             onResult:          { responseAPIResult in
-                                                guard let result = (responseAPIResult as! ResponseAPIRegistrationFirstStepResult).result else {
-                                                    let responseAPIError = (responseAPIResult as! ResponseAPIRegistrationFirstStepResult).error
-                                                    Logger.log(message: "\nAPI `registration.firstStep` response mapping error: \n\(responseAPIError!.message)\n", event: .error)
-                                                    return errorHandling(responseAPIError!)
-                                                }
-                                                
-                                                Logger.log(message: "\nAPI `registration.firstStep` response result: \n\(responseAPIResult)\n", event: .debug)
-                                                
-                                                do {
-                                                    try KeychainManager.save(data: [
-                                                        Config.registrationStepKey: "verify",
-                                                        Config.registrationUserPhoneKey: phone,
-                                                        Config.registrationSmsCodeKey: result.code,
-                                                        Config.registrationSmsNextRetryKey: result.nextSmsRetry
-                                                    ])
-                                                    
-                                                    responseHandling(result)
-                                                } catch {
-                                                    Logger.log(message: error.localizedDescription, event: .error)
-                                                    errorHandling(ResponseAPIError(code: 404, message: error.localizedDescription, currentState: nil))
-                                                }
-        },
-                                             onError: { errorAPI in
-                                                Logger.log(message: "\nAPI `registration.firstStep` response  error: \n\(errorAPI.localizedDescription)\n", event: .error)
-                                                errorHandling(ResponseAPIError(code: Int64(errorAPI.caseInfo.code), message: errorAPI.caseInfo.message, currentState: nil))
-        })
-    }
-
-    // API `registration.verify`
-    public func verify(phone:               String,
-                       code:                UInt64,
-                       responseHandling:    @escaping (ResponseAPIRegistrationVerify) -> Void,
-                       errorHandling:       @escaping (ResponseAPIError) -> Void) {
-        // Offline mode
-        if (!Config.isNetworkAvailable) { return errorHandling(ResponseAPIError(code: 503, message: "No Internet Connection", currentState: nil)) }
-
-        let methodAPIType = MethodAPIType.verify(phone: phone, code: code)
-        
-        Broadcast.instance.executeGETRequest(byContentAPIType:  methodAPIType,
-                                             onResult:          { responseAPIResult in
-                                                guard let result = (responseAPIResult as! ResponseAPIRegistrationVerifyResult).result else {
-                                                    let responseAPIError = (responseAPIResult as! ResponseAPIRegistrationVerifyResult).error
-                                                    Logger.log(message: "\nAPI `registration.verify` response mapping error: \n\(responseAPIError!.message)\n", event: .error)
-                                                    return errorHandling(responseAPIError!)
-                                                }
-                                                
-                                                Logger.log(message: "\nAPI `registration.verify` response result: \n\(responseAPIResult)\n", event: .debug)
-                                                
-                                                do {
-                                                    try KeychainManager.save(data: [Config.registrationStepKey: "setUsername", Config.registrationUserPhoneKey: phone, Config.registrationSmsCodeKey: code])
-                                                    
-                                                    responseHandling(result)
-                                                } catch {
-                                                    Logger.log(message: error.localizedDescription, event: .error)
-                                                    errorHandling(ResponseAPIError(code: 404, message: error.localizedDescription, currentState: nil))
-                                                }
-        },
-                                             onError:           { errorAPI in
-                                                Logger.log(message: "\nAPI `registration.verify` response error: \n\(errorAPI.localizedDescription)\n", event: .error)
-                                                errorHandling(ResponseAPIError(code: Int64(errorAPI.caseInfo.code), message: errorAPI.caseInfo.message, currentState: nil))
-        })
-    }
-    
-    // API `registration.resendSmsCode`
-    public func resendSmsCode(phone:                String,
-                              responseHandling:     @escaping (ResponseAPIResendSmsCode) -> Void,
-                              errorHandling:        @escaping (ErrorAPI) -> Void) {
-        // Offline mode
-        if (!Config.isNetworkAvailable) { return errorHandling(ErrorAPI.disableInternetConnection(message: nil)) }
-        
-        let methodAPIType = MethodAPIType.resendSmsCode(phone: phone, isDebugMode: isDebugMode)
-        
-        Broadcast.instance.executeGETRequest(byContentAPIType:  methodAPIType,
-                                             onResult:          { responseAPIResult in
-                                                Logger.log(message: "\nresponse API Result = \(responseAPIResult)\n", event: .debug)
-                                                
-                                                guard let result = (responseAPIResult as! ResponseAPIResendSmsCodeResult).result else {
-                                                    let responseAPIError = (responseAPIResult as! ResponseAPIResendSmsCodeResult).error
-                                                    Logger.log(message: "\nAPI `registration.resendSmsCode` response mapping error: \n\(responseAPIError!.message)\n", event: .error)
-                                                    return errorHandling(ErrorAPI.jsonParsingFailure(message: "\(responseAPIError!.message)"))
-                                                }
-                                                
-                                                Logger.log(message: "\nAPI `registration.resendSmsCode` response result: \n\(responseAPIResult)\n", event: .debug)
-                                                
-                                                do {
-                                                    try KeychainManager.save(data: [Config.registrationStepKey: "verify", Config.registrationUserPhoneKey: phone, Config.registrationSmsCodeKey: result.code, Config.registrationSmsNextRetryKey: result.nextSmsRetry])
-                                                    
-                                                    responseHandling(result)
-                                                } catch {
-                                                    Logger.log(message: error.localizedDescription, event: .error)
-                                                    errorHandling(ErrorAPI.responseUnsuccessful(message: error.localizedDescription))
-                                                }
-        },
-                                             onError:           { errorAPI in
-                                                Logger.log(message: "\nAPI `registration.resendSmsCode` response error: \n\(errorAPI.localizedDescription)\n", event: .error)
-                                                errorHandling(errorAPI)
-        })
-    }
-    
-    // API `registration.setUsername`
-    public func setUser(id:                     String,
-                        phone:                  String,
-                        responseHandling:       @escaping (ResponseAPIRegistrationSetUsername) -> Void,
-                        errorHandling:          @escaping (ErrorAPI) -> Void) {
-        // Offline mode
-        if (!Config.isNetworkAvailable) { return errorHandling(ErrorAPI.disableInternetConnection(message: nil)) }
-        
-        let methodAPIType = MethodAPIType.setUser(id: id, phone: phone)
-
-        Broadcast.instance.executeGETRequest(byContentAPIType:  methodAPIType,
-                                             onResult:          { responseAPIResult in
-                                                guard let result = (responseAPIResult as! ResponseAPIRegistrationSetUsernameResult).result else {
-                                                    let responseAPIError = (responseAPIResult as! ResponseAPIRegistrationSetUsernameResult).error
-                                                    Logger.log(message: "\nAPI `registration.setUsername` response mapping error: \n\(responseAPIError!.message)\n", event: .error)
-                                                    return errorHandling(ErrorAPI.jsonParsingFailure(message: "\(responseAPIError!.message)"))
-                                                }
-                                                
-                                                Logger.log(message: "\nAPI `registration.setUsername` response result: \n\(responseAPIResult)\n", event: .debug)
-                                                
-                                                do {
-                                                    try KeychainManager.save(data: [Config.registrationStepKey: "toBlockChain", Config.registrationUserPhoneKey: phone, Config.registrationUserIDKey: id])
-                                                    
-                                                    responseHandling(result)
-                                                } catch {
-                                                    Logger.log(message: error.localizedDescription, event: .error)
-                                                    errorHandling(ErrorAPI.responseUnsuccessful(message: error.localizedDescription))
-                                                }
-        },
-                                             onError:           { errorAPI in
-                                                Logger.log(message: "\nAPI `registration.setUsername` response error: \n\(errorAPI.localizedDescription)\n", event: .error)
-                                                errorHandling(errorAPI)
-        })
-    }
-    
-    // API `registration.toBlockChain`
-    public func toBlockChain(id:                    String,
-                             phone:                 String,
-                             responseHandling:      @escaping (Bool) -> Void,
-                             errorHandling:         @escaping (ErrorAPI) -> Void) {
-        // Offline mode
-        if (!Config.isNetworkAvailable) { return errorHandling(ErrorAPI.disableInternetConnection(message: nil)) }
-        
-        let userkeys = generateKeys(userID: id, password: String.randomString(length: 12))
-        
-        let methodAPIType = MethodAPIType.toBlockChain(userID: id, keys: userkeys)
-        
-        Broadcast.instance.executeGETRequest(byContentAPIType:  methodAPIType,
-                                             onResult:          { responseAPIResult in
-                                                guard let chainResult = (responseAPIResult as? ResponseAPIRegistrationToBlockChainResult)?.result else {
-                                                    let responseAPIError = (responseAPIResult as! ResponseAPIRegistrationToBlockChainResult).error
-                                                    Logger.log(message: "\nAPI `registration.toBlockChain` response mapping error: \n\(responseAPIError!.message)\n", event: .error)
-                                                    return errorHandling(ErrorAPI.jsonParsingFailure(message: "\(responseAPIError!.message)"))
-                                                }
-                                                
-                                                // Save in Keychain
-                                                Logger.log(message: "\nAPI `registration.toBlockChain` response result: \n\(responseAPIResult)\n", event: .debug)
-                                                
-                                                do {
-                                                    try KeychainManager.save(data: [
-                                                        Config.registrationStepKey:            "firstStep",
-                                                        Config.registrationUserNameKey:        chainResult.username,
-                                                        Config.registrationUserIDKey:          chainResult.userId,
-                                                        Config.currentUserPublicOwnerKey:      userkeys["owner"]!.publicKey,
-                                                        Config.currentUserPrivateOwnerKey:     userkeys["owner"]!.privateKey,
-                                                        Config.currentUserPublicActiveKey:     userkeys["active"]!.publicKey,
-                                                        Config.currentUserPrivateActiveKey:    userkeys["active"]!.privateKey,
-                                                        Config.currentUserPublicPostingKey:    userkeys["posting"]!.publicKey,
-                                                        Config.currentUserPrivatePostingKey:   userkeys["posting"]!.privateKey,
-                                                        Config.currentUserPublickMemoKey:      userkeys["memo"]!.publicKey,
-                                                        Config.currentUserPrivateMemoKey:      userkeys["memo"]!.privateKey
-                                                        ])
-                                                    
-                                                    PDFManager.createPDFFile()
-                                                    
-                                                    // Auth new account
-                                                    DispatchQueue.main.async {
-                                                        RestAPIManager.instance.authorize(
-                                                            userID: Config.currentUser!.id!,
-                                                            userActiveKey: Config.currentUser!.activeKey!,
-                                                            responseHandling: { response in
-                                                                responseHandling(true)
-                                                            },
-                                                            errorHandling: { error in
-                                                                errorHandling(error.toErrorAPI())
-                                                            })
-                                                    }
-                                                    
-                                                    
-                                                } catch {
-                                                    Logger.log(message: error.localizedDescription, event: .error)
-                                                    errorHandling(ErrorAPI.responseUnsuccessful(message: error.localizedDescription))
-                                                }
-                                                
-        },
-                                             onError: { errorAPI in
-                                                Logger.log(message: "\nAPI `registration.toBlockChain` response error: \n\(errorAPI.localizedDescription)\n", event: .error)
-                                                errorHandling(errorAPI)
-        })
-    }
-    
-    
     //  MARK: - Contract `gls.social`
     /// Posting image
     public func posting(image:              UIImage,
